added test comments
